requirement.txt
Django==2.1.3
mysqlclient==1.3.13
django-rest-framework==0.1.0